+++ bosh-setup/scripts/setup_env.sh
@@ -12,7 +12,7 @@ set -e
 
 tenant_id=$1
 client_id=$2
-client_secret=$(echo $3 | base64 --decode)
+base64_encoded_client_secret_or_certificate=$3
 custom_data_file="/var/lib/cloud/instance/user-data.txt"
 settings=$(cat ${custom_data_file})
 
@@ -30,7 +30,12 @@ function install_bosh_cli() {
   sudo mv ./bosh-cli-* /usr/local/bin/bosh
 }
 
+function client_secret_or_certificate() {
+  echo ${base64_encoded_client_secret_or_certificate} | base64 --decode
+}
+
 environment=$(get_setting ENVIRONMENT)
+service_principal_type=$(get_setting SERVICE_PRINCIPAL_TYPE)
 
 set +e
 
@@ -63,6 +68,46 @@ home_dir="/home/$username"
 manifests_dir="$home_dir/example_manifests"
 mkdir -p $manifests_dir
 cp *.yml $manifests_dir
+pushd $manifests_dir > /dev/null
+  # Enable availability zones if needed
+  use_availability_zones=$(get_setting USE_AVAILABILITY_ZONES)
+  if [ "$use_availability_zones" == "enabled" ]; then
+    sed -i '1,5d' cloud-config.yml
+    cat - cloud-config.yml > cloud-config-azs-enabled.yml << EOF
+---
+azs:
+- name: z1
+  cloud_properties:
+    availability_zone: '1'
+- name: z2
+  cloud_properties:
+    availability_zone: '2'
+- name: z3
+  cloud_properties:
+    availability_zone: '3'
+EOF
+    mv cloud-config-azs-enabled.yml cloud-config.yml
+  fi
+  if [ "${service_principal_type}" == "Certificate" ]; then
+    cat > service-principal-certificate.yml << EOF
+certificate: |-
+$(client_secret_or_certificate | sed 's/^/  /')
+EOF
+  fi
+  if [ "$environment" = "AzureStack" ]; then
+    if [ "$(get_setting AZURE_STACK_CA_ROOT_CERTIFICATE | base64 --decode)" = "" ]; then
+      cat > azure-stack-ca-cert.yml << EOF
+ca_cert: |-
+$(cat /var/lib/waagent/Certificates.pem | sed 's/^/  /')
+EOF
+    else
+      cat > azure-stack-ca-cert.yml << EOF
+ca_cert: |-
+$(get_setting AZURE_STACK_CA_ROOT_CERTIFICATE | base64 --decode | sed 's/^/  /')
+EOF
+    fi
+  fi
+popd  > /dev/null
 chmod 775 $manifests_dir
 chmod 644 $manifests_dir/*
 dpkg -i cf-cli*
@@ -79,46 +124,58 @@ bosh create-env ~/example_manifests/bosh.yml \\
   --state=state.json \\
   --vars-store=~/bosh-deployment-vars.yml \\
   -o ~/example_manifests/cpi.yml \\
+  -o ~/example_manifests/custom-cpi-release.yml \\
   -o ~/example_manifests/custom-environment.yml \\
   -o ~/example_manifests/use-azure-dns.yml \\
   -o ~/example_manifests/jumpbox-user.yml \\
+  -v director_name=azure \\
+  -v internal_cidr=10.0.0.0/24 \\
+  -v internal_gw=10.0.0.1 \\
+  -v internal_ip=10.0.0.4 \\
+  -v cpi_release_url=$(get_setting BOSH_AZURE_CPI_RELEASE_URL) \\
+  -v cpi_release_sha1=$(get_setting BOSH_AZURE_CPI_RELEASE_SHA1) \\
+  -v director_vm_instance_type=$(get_setting BOSH_VM_SIZE) \\
+  -v resource_group_name=$(get_setting RESOURCE_GROUP_NAME) \\
+  -v vnet_name=$(get_setting VNET_NAME) \\
+  -v subnet_name=$(get_setting SUBNET_NAME_FOR_BOSH) \\
+  -v default_security_group=$(get_setting NSG_NAME_FOR_BOSH) \\
+  -v environment=$(get_setting ENVIRONMENT) \\
+  -v subscription_id=$(get_setting SUBSCRIPTION_ID) \\
+  -v tenant_id=${tenant_id} \\
+  -v client_id=${client_id} \\
 EOF
+
+if [ "${service_principal_type}" == "Password" ]; then
+  cat >> "$home_dir/deploy_bosh.sh" << EOF
+  -v client_secret="$(client_secret_or_certificate)" \\
+EOF
+elif [ "${service_principal_type}" == "Certificate" ]; then
+  cat >> "$home_dir/deploy_bosh.sh" << EOF
+  -o ~/example_manifests/use-service-principal-with-certificate.yml \\
+  -l ~/example_manifests/service-principal-certificate.yml \\
+EOF
+fi
 
 if [ "$environment" = "AzureChinaCloud" ]; then
   cat >> "$home_dir/deploy_bosh.sh" << EOF
   -o ~/example_manifests/use-managed-disks.yml \\
   -o ~/example_manifests/use-mirror-releases-for-bosh.yml \\
-  -o ~/example_manifests/custom-ntp-server.yml \\
+  -o ~/example_manifests/custom-ntp-server.yml
 EOF
 elif [ "$environment" = "AzureStack" ]; then
   cat >> "$home_dir/deploy_bosh.sh" << EOF
-  -o ~/example_manifests/azure-stack-properties.yml \\
   -v storage_account_name=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME) \\
+  -o ~/example_manifests/azure-stack-properties.yml \\
   -v azure_stack_domain=$(get_setting AZURE_STACK_DOMAIN) \\
   -v azure_stack_resource=$(get_setting AZURE_STACK_RESOURCE) \\
+  -v azure_stack_authentication=$(get_setting AZURE_STACK_AUTHENTICATION) \\
+  -l ~/example_manifests/azure-stack-ca-cert.yml
 EOF
 else
   cat >> "$home_dir/deploy_bosh.sh" << EOF
-  -o ~/example_manifests/use-managed-disks.yml \\
+  -o ~/example_manifests/use-managed-disks.yml
 EOF
 fi
-
-cat >> "$home_dir/deploy_bosh.sh" << EOF
-  -v director_name=azure \\
-  -v internal_cidr=10.0.0.0/24 \\
-  -v internal_gw=10.0.0.1 \\
-  -v internal_ip=10.0.0.4 \\
-  -v director_vm_instance_type=$(get_setting BOSH_VM_SIZE) \\
-  -v vnet_name=$(get_setting VNET_NAME) \\
-  -v subnet_name=$(get_setting SUBNET_NAME_FOR_BOSH) \\
-  -v environment=$(get_setting ENVIRONMENT) \\
-  -v subscription_id=$(get_setting SUBSCRIPTION_ID) \\
-  -v tenant_id=${tenant_id} \\
-  -v client_id=${client_id} \\
-  -v client_secret="${client_secret}" \\
-  -v resource_group_name=$(get_setting RESOURCE_GROUP_NAME) \\
-  -v default_security_group=$(get_setting NSG_NAME_FOR_BOSH)
-EOF
 chmod 777 $home_dir/deploy_bosh.sh
 
 cat > "$home_dir/deploy_cloud_foundry.sh" << EOF
@@ -145,12 +202,12 @@ cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
 bosh -n -d cf deploy ~/example_manifests/cf-deployment.yml \\
   --vars-store=~/cf-deployment-vars.yml \\
   -o ~/example_manifests/azure.yml \\
+  -o ~/example_manifests/scale-to-one-az.yml \\
 EOF
 if [ "$environment" = "AzureChinaCloud" ]; then
   cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
   -o ~/example_manifests/use-azure-storage-blobstore.yml \\
   -o ~/example_manifests/use-mirror-releases-for-cf.yml \\
-  -o ~/example_manifests/scale-to-one-az.yml \\
   -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io \\
   -v environment=$(get_setting ENVIRONMENT) \\
   -v blobstore_storage_account_name=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME) \\
@@ -162,14 +219,11 @@ if [ "$environment" = "AzureChinaCloud" ]; then
 EOF
 elif [ "$environment" = "AzureStack" ]; then
   cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
-  -o ~/example_manifests/scale-to-one-az.yml \\
-  -o ~/example_manifests/scale-to-availability-set-no-HA.yml \\
   -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io
 EOF
 else
   cat >> "$home_dir/deploy_cloud_foundry.sh" << EOF
   -o ~/example_manifests/use-azure-storage-blobstore.yml \\
-  -o ~/example_manifests/scale-to-one-az.yml \\
   -v system_domain=$(get_setting CLOUD_FOUNDRY_PUBLIC_IP).xip.io \\
   -v environment=$(get_setting ENVIRONMENT) \\
   -v blobstore_storage_account_name=$(get_setting DEFAULT_STORAGE_ACCOUNT_NAME) \\