+++ bosh-setup/CHANGELOG.md
@@ -1,3 +1,17 @@
+# v3.4.0 (2018-01-30)
+
+- Bump cf-deployment to [`v1.9.0`](https://github.com/cloudfoundry/cf-deployment/tree/v1.9.0).
+- Bump bosh release to `264.7.0`
+- Bump stemcell in bosh-deployment to `3468.21`
+- Bump stemcell in cf-deployment to `3468.19`
+
+# v3.3.0 (2018-01-09)
+
+- Upgrade Azure CPI version to v35.0.0. Please see new features in [bosh-azure-cpi-release v35.0.0](https://github.com/cloudfoundry-incubator/bosh-azure-cpi-release/releases/tag/v35.0.0)
+- Bump cf-deployment to [`v1.6.0`](https://github.com/cloudfoundry/cf-deployment/tree/v1.6.0).
+- Support [Availability Zones](https://docs.microsoft.com/en-us/azure/availability-zones/az-overview).
+- Set the default value of `autoDeployCloudFoundry` to `disabled`
+
 # v3.2.0 (2017-12-20)
 
 - Bump [bosh-deployment](https://github.com/cloudfoundry/bosh-deployment/blob/218e6d5030d89ca9f31c50b8b308e2a78d2a0997/bosh.yml) and [cf-deployment](https://github.com/cloudfoundry/cf-deployment/blob/v1.4.0/cf-deployment.yml).